+++ oow.js
@@ -1,3 +1,4 @@
 'use strict';
 
 require('./src/collection_extensions');
+require('./src/date_extensions');